+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Next (TBD) - Master
 
 * add custom `url_for` method in TilerFactory to retrieve `prefixed` endpoint URL (#95)
+* remove magic `titiler.dependencies.PathParams` mosaicid path translation, where a user could pass `url=mosaicid://` to the endpoint.
 
 ## 0.1.0-alpha.2 (2020-09-01)